add pl.cut utility
