+++ py-polars/docs/source/reference/functions.rst
@@ -30,6 +30,7 @@ Eager/Lazy functions
 
    arg_where
    concat
+   cut
    date_range
    get_dummies
    repeat
